+++ src/Middleware/HostnameActions.php
@@ -54,6 +54,8 @@ class HostnameActions
         $hostname = app(CurrentHostname::class);
 
         if ($hostname != null) {
+            $this->setAppUrl($request, $hostname);
+
             if ($hostname->under_maintenance_since) {
                 return $this->maintenance($hostname);
             }
@@ -113,6 +115,21 @@ class HostnameActions
         if (config('tenancy.hostname.abort-without-identified-hostname')) {
             $this->emitEvent(new NoneFound($request));
             return abort(404);
+        }
+    }
+
+    /**
+     * Forces the app.url configuration to the tenant hostname domain.
+     *
+     * @param Request  $request
+     * @param Hostname $hostname
+     */
+    protected function setAppUrl(Request $request, Hostname $hostname)
+    {
+        if (config('tenancy.hostname.update-app-url', false)) {
+            config([
+                'app.url' => sprintf('%s://', $request->getScheme(), $hostname->fqdn)
+            ]);
         }
     }
 }