+++ src/Traits/MutatesSeedCommands.php
@@ -51,10 +51,6 @@ trait MutatesSeedCommands
         $this->input->setOption('force', true);
         $this->input->setOption('database', $this->connection->tenantName());
 
-        if (! $this->option('class')) {
-            $this->input->setOption('class', config('tenancy.db.tenant-seed-class'));
-        }
-
         $this->processHandle(function (Website $website) {
             $this->connection->set($website);
 
@@ -71,8 +67,14 @@ trait MutatesSeedCommands
      */
     protected function getOptions()
     {
-        return array_merge([
+        foreach ($options = parent::getOptions() as $option) {
+            if ($option[0] === 'class') {
+                $option[4] = config('tenancy.db.tenant-seed-class', false) ?: $option[4];
+            }
+        }
+
+        return array_merge($options, [
             $this->addWebsiteOption()
-        ], parent::getOptions());
+        ]);
     }
 }